Auth_app/requirement.txt
Flask==2.3.2
Authlib==1.2.0
SQLAlchemy==2.0.19
pymysql==1.0.3
python-dotenv==1.0.0
